requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
html2text>=2020.1.16
openai>=1.0.0
python-telegram-bot>=20.0